requirements.txt
Flask>=2.2.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.0
Flask-Bcrypt>=1.0.1
python-dotenv>=0.21.0
anthropic>=0.34.0
Flask-Cors>=3.0.10
psycopg2-binary>=2.9.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
